p2p:move old p2p to gossip directory
