+++ p2p/gossip/doc.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package p2p
+package gossip
 
 //p2p 网络模块：
 //主要的功能是实现区块链数据的广播，交易的广播的功能。
