+++ p2p/gossip/net_test.go
@@ -1,7 +1,7 @@
 // Copyright Fuzamei Corp. 2018 All Rights Reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
-package p2p
+package gossip
 
 import (
 	"net"
