+++ p2p/gossip/p2p.go
@@ -3,7 +3,7 @@
 // license that can be found in the LICENSE file.
 
 // Package p2p 实现了chain33网络协议
-package p2p
+package gossip
 
 import (
 	"fmt"
