+++ p2p/gossip/p2p_test.go
@@ -1,4 +1,4 @@
-package p2p
+package gossip
 
 import (
 	"encoding/hex"
